Adding urlview-compat optional argument
This adds a switch '-u/--urlview-compat' which changes the behavior of
what urlscan expects from stdin or the 'message' argument.  Instead of
expecting a filename for an email-like file, it instead expects any text
that may or may not contain urls.  It then puts the text in an
email-like temporary file and passes it on to normal urlscan logic.

The goal of this change is to allow for more urlview-esq behavior.

Example:
    echo "http://www.google.com\nhttp://www.yahoo.com" | urlview

The above example will extract the two urls when used with urlview, but
urlscan does not find any urls because it is expecting a filename.