Allow $ as an acceptable trailing character
My company has recently started using a URL defense service that
encapsulates email URLs and routes them to a URL filtering service.
These URLs also include a checksum that always ends with a $ character.

urlscan does not currently recognize the $ as part of a link and
therefore removes it, invalidating the checksum and giving me an error
in the URL defense service. By including the $ as an acceptable
character, I can now successfully use email links with mutt and the new
service.